requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-compat==1.0.15
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-rq==2.8.1
django-widget-tweaks==1.4.12
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
Pillow==10.0.0
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
rq==1.15.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.3
